Move offline rpgo params and submap align params to params submodule
Previously were located in offline_rpgo and submap_align modules
respectively. Small clean up.
