+++ demo/params/demo/offline_rpgo.yaml
@@ -4,3 +4,4 @@ odom_t_std: 0.02 # 2 cm
 odom_r_std: 0.001745 # .1 deg
 lc_t_std: 2.0 # 2 m
 lc_r_std: 0.1745 # 10 deg
+sparsified: False
